[refactor] :recycle: move environment to constants (#276)
