+++ src/config/database.ts
@@ -1,7 +1,7 @@
 import mongoose from 'mongoose';
 
+import { MONGO_URI } from './constants';
 import { Logger } from '../helpers/loggerHelper';
-import { MONGO_URI } from '../constants/environment';
 
 /**
  * Connects to the MongoDB database using the provided URI or a default local URI.
