+++ src/constants/blockchain.ts
@@ -1,8 +1,2 @@
-import { GCP_BUCKET_BASE_URL } from './environment';
-
-export const defaultNftImage = `${GCP_BUCKET_BASE_URL}/images/default_nft.png`;
-
-
-
 export const LIFI_SLIPPAGE = 30 / 1000;
 export const LIFI_TYPE = 'SAFEST';
