+++ src/controllers/tokenController.ts
@@ -3,7 +3,7 @@ import { FastifyReply, FastifyRequest, FastifyInstance } from 'fastify';
 
 import Token, { IToken } from '../models/token';
 import { Logger } from '../helpers/loggerHelper';
-import { BUN_ENV, SIGNING_KEY } from '../constants/environment';
+import { BUN_ENV, SIGNING_KEY } from '../config/constants';
 import { returnErrorResponse, returnSuccessResponse } from '../helpers/responseFormatterHelper';
 
 /**
