+++ src/services/bucketService.ts
@@ -2,8 +2,8 @@ import axios from 'axios';
 import { ethers } from 'ethers';
 import NodeCache from 'node-cache';
 
+import { GCP_ABIs } from '../config/constants';
 import { Logger } from '../helpers/loggerHelper';
-import { GCP_ABIs } from '../constants/environment';
 
 export type ABI = ethers.ContractInterface;
