+++ src/services/paymasterService.ts
@@ -3,7 +3,7 @@ import { ethers } from 'ethers';
 import { Logger } from '../helpers/loggerHelper';
 import { PackedUserOperationType } from '../types/userOperation';
 import { createPaymasterAndData } from '../helpers/paymasterHelper';
-import { PAYMASTER_MIN_BALANCE, PAYMASTER_TARGET_BALANCE } from '../constants/environment';
+import { PAYMASTER_MIN_BALANCE, PAYMASTER_TARGET_BALANCE } from '../config/constants';
 
 /**
  * Adds paymaster data to a UserOperation.
