+++ src/services/userOperationService.ts
@@ -9,7 +9,7 @@ import {
   PRE_VERIFICATION_GAS,
   VERIFICATION_GAS_LIMIT,
   MAX_PRIORITY_FEE_PER_GAS
-} from '../constants/environment';
+} from '../config/constants';
 
 /**
  * Creates a generic user operation for any type of transaction.
